ghana_decides/requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
Pillow
pytz
sqlparse
psycopg2-binary
redis
celery
channels-redis
requests
pyfcm
django-cors-headers
daphne
django-extensions
django-ratelimit
prometheus-client
Werkzeug
